cbs/validator/requirements.txt
ansicolors~=1.1
boto3~=1.34
urllib3>=1.25.4,<1.27 # https://github.com/boto/botocore/issues/3138#issuecomment-1986927748
